ERA-11293: Rename scripts and config folders and fix eslint parser babel-preset-react-app
