+++ react-app-config/modules.js
