+++ react-app-config/webpack/persistentCache/createEnvironmentHash.js
