+++ react-scripts/start.js
@@ -12,7 +12,7 @@ process.on('unhandledRejection', err => {
 });
 
 // Ensure environment variables are read.
-require('../config/env');
+require('../react-app-config/env');
 
 const fs = require('fs');
 const chalk = require('react-dev-utils/chalk');
@@ -28,10 +28,10 @@ const {
 } = require('react-dev-utils/WebpackDevServerUtils');
 const openBrowser = require('react-dev-utils/openBrowser');
 const semver = require('semver');
-const paths = require('../config/paths');
-const configFactory = require('../config/webpack.config');
-const createDevServerConfig = require('../config/webpackDevServer.config');
-const getClientEnvironment = require('../config/env');
+const paths = require('../react-app-config/paths');
+const configFactory = require('../react-app-config/webpack.config');
+const createDevServerConfig = require('../react-app-config/webpackDevServer.config');
+const getClientEnvironment = require('../react-app-config/env');
 const react = require(require.resolve('react', { paths: [paths.appPath] }));
 
 const env = getClientEnvironment(paths.publicUrlOrPath.slice(0, -1));
