+++ react-scripts/test.js
@@ -13,7 +13,7 @@ process.on('unhandledRejection', err => {
 });
 
 // Ensure environment variables are read.
-require('../config/env');
+require('../react-app-config/env');
 
 const jest = require('jest');
 const execSync = require('child_process').execSync;
